refactor: reorganize files
Flattens src file structure and extracts config to new file.
