+++ src/blockConfig.js
@@ -1,0 +1,17 @@
+const configSchema = ElementPropTypes => {
+    return {
+        color: ElementPropTypes.shape({
+            background: ElementPropTypes.color
+        }),
+        text: ElementPropTypes.string
+    };
+};
+
+const defaultConfig = {
+    color: {
+        background: 'transparent'
+    },
+    text: 'Default prop'
+};
+
+export { configSchema, defaultConfig };
